update ui and backend

+++ dotnet-backend/BattleArena.Domain/Models/Player.cs
@@ -1,4 +1,4 @@
 namespace BattleArena.Domain.Models;
 
-public sealed record Player(int Id, string UserName, int Wins, int Losses, int Wager, int PowerIndex);
+public sealed record Player(long Id, string UserName, int Wins, int Losses, int Wager, int PowerIndex);
